// blockchain/typechain-types/TaskManager.ts
export declare namespace TaskManager {
  export type TaskStruct = {
    id: BigNumberish;
    content: string;
    completed: boolean;
    owner: AddressLike;
  };

  export type TaskStructOutput = [
    id: bigint,
    content: string,
    completed: boolean,
    owner: string
  ] & { id: bigint; content: string; completed: boolean; owner: string };
}

export interface TaskManagerInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "addTask"
      | "deleteTask"
      | "getMyTasks"
      | "toggleTaskComplete"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "TaskAdded" | "TaskDeleted" | "TaskToggled"
  ): EventFragment;

  encodeFunctionData(functionFragment: "addTask", values: [string]): string;
  encodeFunctionData(
    functionFragment: "deleteTask",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getMyTasks",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "toggleTaskComplete",
    values: [BigNumberish]
  ): string;

  decodeFunctionResult(functionFragment: "addTask", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "deleteTask", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getMyTasks", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "toggleTaskComplete",
    data: BytesLike
  ): Result;
}

export namespace TaskAddedEvent {
  export type InputTuple = [
    taskId: BigNumberish,
    owner: AddressLike,
    content: string
  ];
  export type OutputTuple = [taskId: bigint, owner: string, content: string];
  export interface OutputObject {
    taskId: bigint;
    owner: string;
    content: string;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace TaskDeletedEvent {
  export type InputTuple = [taskId: BigNumberish];
  export type OutputTuple = [taskId: bigint];
  export interface OutputObject {
    taskId: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace TaskToggledEvent {
  export type InputTuple = [taskId: BigNumberish, completed: boolean];
  export type OutputTuple = [taskId: bigint, completed: boolean];
  export interface OutputObject {
    taskId: bigint;
    completed: boolean;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface TaskManager extends BaseContract {
  connect(runner?: ContractRunner | null): TaskManager;
  waitForDeployment(): Promise<this>;

  interface: TaskManagerInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  addTask: TypedContractMethod<[content: string], [void], "nonpayable">;

  deleteTask: TypedContractMethod<[taskId: BigNumberish], [void], "nonpayable">;

  getMyTasks: TypedContractMethod<[], [TaskManager.TaskStructOutput[]], "view">;

  toggleTaskComplete: TypedContractMethod<
    [taskId: BigNumberish],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "addTask"
  ): TypedContractMethod<[content: string], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "deleteTask"
  ): TypedContractMethod<[taskId: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "getMyTasks"
  ): TypedContractMethod<[], [TaskManager.TaskStructOutput[]], "view">;
  getFunction(
    nameOrSignature: "toggleTaskComplete"
  ): TypedContractMethod<[taskId: BigNumberish], [void], "nonpayable">;

  getEvent(
    key: "TaskAdded"
  ): TypedContractEvent<
    TaskAddedEvent.InputTuple,
    TaskAddedEvent.OutputTuple,
    TaskAddedEvent.OutputObject
  >;
  getEvent(
    key: "TaskDeleted"
  ): TypedContractEvent<
    TaskDeletedEvent.InputTuple,
    TaskDeletedEvent.OutputTuple,
    TaskDeletedEvent.OutputObject
  >;
  getEvent(
    key: "TaskToggled"
  ): TypedContractEvent<
    TaskToggledEvent.InputTuple,
    TaskToggledEvent.OutputTuple,
    TaskToggledEvent.OutputObject
  >;

  filters: {
    "TaskAdded(uint256,address,string)": TypedContractEvent<
      TaskAddedEvent.InputTuple,
      TaskAddedEvent.OutputTuple,
      TaskAddedEvent.OutputObject
    >;
    TaskAdded: TypedContractEvent<
      TaskAddedEvent.InputTuple,
      TaskAddedEvent.OutputTuple,
      TaskAddedEvent.OutputObject
    >;

    "TaskDeleted(uint256)": TypedContractEvent<
      TaskDeletedEvent.InputTuple,
      TaskDeletedEvent.OutputTuple,
      TaskDeletedEvent.OutputObject
    >;
    TaskDeleted: TypedContractEvent<
      TaskDeletedEvent.InputTuple,
      TaskDeletedEvent.OutputTuple,
      TaskDeletedEvent.OutputObject
    >;

    "TaskToggled(uint256,bool)": TypedContractEvent<
      TaskToggledEvent.InputTuple,
      TaskToggledEvent.OutputTuple,
      TaskToggledEvent.OutputObject
    >;
    TaskToggled: TypedContractEvent<
      TaskToggledEvent.InputTuple,
      TaskToggledEvent.OutputTuple,
      TaskToggledEvent.OutputObject
    >;
  };
}